added stuff

+++ src/main/java/com/codecool/DAO/HeroRepository.java
@@ -8,7 +8,8 @@ import java.util.List;
 
 public interface HeroRepository extends CrudRepository<Hero, String>{
 
-    List<Hero> findByUserId(Long userId);
-    Hero findById(Long id);
-    void deleteById(Long id);
+    List<Hero> findByUserId(int userId);
+
+    Hero findById(int id);
+    void deleteById(int id);
 }
